fix typing, imports, docstrings, examples
